http://ehc.ac/p/epydoc/bugs/356/
There is a 'bug' of sorts in epydoc's generated 'epydoc.js' file. Browsers implement the 'onmouseout' event in a strange way that results in any link clicking auto-killing the doclink when hovered. There is a fix for that described here:
http://www.quirksmode.org/js/events_mouse.html#mouseover
I've uploaded a patch file that implements the above technique, applying it to epydoc/docwriter/html_colorize.py appears to fix the issue in all browsers I tested.